fix: stack screen issue so it displays with proper formatting
