+++ App.js
@@ -1,26 +1,26 @@
 import { useState, useEffect } from 'react';
+import "react-native-gesture-handler"
 import { StyleSheet, SafeAreaView, View, StatusBar } from 'react-native';
 import ViewProblem from './app/Pages/ViewProblem';
 import CreateProblem from './app/Pages/CreateProblem';
 import Home from './app/Pages/Home';
 import MenuBar from './app/Components/MenuBar';
-import { getUserWalls } from './app/Components/apiCalls';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import ViewAllProblems from '../klime_fe/app/Pages/ViewAllProblems'
 
 
 export default function App() {
-  const [savedWalls, setSavedWalls] = useState([])
+  // const [savedWalls, setSavedWalls] = useState([])
   const [vectorColor, setVectorColor] = useState('60FF46')
 
-  useEffect(() => {
-    (async () => {
-      const data = await getUserWalls();
-      const modifiedData = data.data.map((data) => ({ id: data.id, ...data.attributes }))
-      setSavedWalls(modifiedData)
-    })();
-  }, [])
+  // useEffect(() => {
+  //   (async () => {
+  //     const data = await getUserWalls();
+  //     const modifiedData = data.data.map((data) => ({ id: data.id, ...data.attributes }))
+  //     setSavedWalls(modifiedData)
+  //   })();
+  // }, [])
 
   const handleVectorColor = (color) => {
     switch (color) {
@@ -37,15 +37,17 @@ export default function App() {
   }
   const Stack = createNativeStackNavigator()
 
+
   return (
-    <NavigationContainer>
+    <NavigationContainer style={{ ...styles.container, ...styles.androidSafeArea }}>
       <Stack.Navigator>
-        <Stack.Screen name='View All Problems' component={ViewAllProblems} />
-      </Stack.Navigator>
-      <SafeAreaView style={{ ...styles.container, ...styles.androidSafeArea }}>
+          <Stack.Screen name='Home' component={Home} />
+          <Stack.Screen name='View All Problems' component={ViewAllProblems} />
+          <Stack.Screen name='View Problem' component={ViewProblem} />
+        </Stack.Navigator>
+      <SafeAreaView>
         <StatusBar />
         <View style={styles.contentContainer}>
-          <Home savedWalls={savedWalls} />
           {/* <ViewProblem /> */}
           {/* <CreateProblem vectorColor={vectorColor} />  */}
         </View>
