+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@react-navigation/native": "^6.1.7",
+    "@react-navigation/native-stack": "^6.9.13",
     "expo": "~49.0.8",
     "expo-constants": "~14.4.2",
     "expo-linking": "~5.0.2",
